feat: moved sidebar into components
